Define the make_customization_applicator function
This function centralizes the customization application logic.
This logic also includes KeyError handling and written value tracing.
It returns a function which can be used as a callback for traverse.
To create applicator functions, it needs some parameters as well as a
function which returns the value that will be written to the game.